requirements.txt
beautifulsoup4==4.11.1
billboard.py==7.0.0
Flask==2.1.2
flask_login
flask_sqlalchemy
lyricsgenius==3.0.1
Markdown==3.3.7
praw==7.6.0
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.1
Werkzeug==2.1.2
gunicorn